Remove haml and slim tempaltes
It is hard to maintain multiple types of templates especially when you are not familiar with haml or slim, but it only takes a minute or even less to conevrt erb tempaltes to haml or slim. I don't think we can justify the reason to maintain haml and slim templates in the repository.